removing and adding of source

+++ MammoExpert/MammoExpert.PatientServices.Demo/ViewModel/SourcesWindowViewModel.cs
@@ -5,25 +5,37 @@ using MammoExpert.PatientServices.Sources;
 using MammoExpert.PatientServices.PresenterCore;
 using MammoExpert.PatientServices.UI.Controls.ViewModel;
 using System;
+using System.Collections.ObjectModel;
+using System.Collections.Specialized;
 using System.Linq;
+using System.Windows;
 
 namespace MammoExpert.PatientServices.Demo.ViewModel
 {
     public class SourcesWindowViewModel : MainWindowViewModel
     {
+        #region Fields
+
         private List<SourceType> _sourceTypeOptions;
-        private List<Source> _sources;
+        private ObservableCollection<Source> _sources;
+
+        #endregion // Fields
+
+        #region Constructor
 
         public SourcesWindowViewModel()
         {
-
             base.DisplayName = Resources.SourcesWindowViewModel_DisplayName;
 
-            LoadAllSources();
+            _sources = new ObservableCollection<Source>(SourceRepository.GetAllSources());
         }
 
+        #endregion // Constructor
+
+        #region Properties
+
         // список отображаемых источников
-        public List<Source> Sources
+        public ObservableCollection<Source> Sources
         {
             get { return _sources; }
             set
@@ -47,6 +59,10 @@ namespace MammoExpert.PatientServices.Demo.ViewModel
             }
         }
 
+        #endregion // Properties
+
+        #region Commands
+
         public ICommand AddWorkspaceCommand => new ActionCommand<Source>(AddWorkspace);
 
         public ICommand AddSourceCommand => new ActionCommand<SourceType>(OpenConfigurationWindow);
@@ -57,7 +73,11 @@ namespace MammoExpert.PatientServices.Demo.ViewModel
 
         public ICommand ChangeSourceListByType => new ActionCommand<SourceType>(ChangeSourceList);
 
-        public void AddWorkspace(Source source)
+        #endregion // Commands
+
+        #region Private Methods
+
+        private void AddWorkspace(Source source)
         {
             if (source != null)
             {
@@ -66,37 +86,40 @@ namespace MammoExpert.PatientServices.Demo.ViewModel
             }
         }
 
-        public void EditSource(Source source)
-        {
-            if (source != null) WindowFactory.CreateConfigurationWindow(source);
-        }
-
-        public void DeleteSource(Source source)
-        {
-            if (source != null)
-            {
-                SourceRepository.DeleteSource(source);
-            }
-        }
-
-        public void OpenConfigurationWindow(SourceType type)
+        private void OpenConfigurationWindow(SourceType type)
         {
             WindowFactory.CreateConfigurationWindow(type);
         }
 
-        // метод, загружающий источники всех типов
-        private void LoadAllSources()
+        private void EditSource(Source source)
         {
-            if (SourceRepository != null)
-            {
-                _sources = SourceRepository.GetAllSources();
-            }
+            if (source != null) WindowFactory.CreateConfigurationWindow(source);
         }
 
+        private void DeleteSource(Source source)
+        {
+            if (source != null)
+            {
+                MessageBoxResult result = MessageBox.Show(
+                "Вы уверены, что хотите удалить " + source.Name  + "?",
+                "ВНИМАНИЕ!!!",
+                MessageBoxButton.YesNo,
+                MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes)
+                {
+                    SourceRepository.DeleteSource(source);
+                    ChangeSourceList(source.Type);
+                }
+            }
+        }
+        
         // метод, обновляющий список источников согласно выбранному типу источника
         private void ChangeSourceList(SourceType type)
         {
-            Sources = SourceRepository.GetSourcesByType(type);
+            var collection = SourceRepository.GetSourcesByType(type);
+            Sources = new ObservableCollection<Source>(collection);
         }
+
+        #endregion // Private Methods
     }
 }
